backend/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "Django rest API for an e-commerce application"
authors = ["Sean Welch"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.8.1"
boto3 = "1.34.71"
boto3-stubs = "1.34.74"
botocore = "1.34.71"
botocore-stubs = "1.34.69"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
django = "5.0.3"
django-filter = "24.1"
django-stubs = "4.2.7"
django-stubs-ext = "4.2.7"
djangorestframework = "3.15.1"
djangorestframework-stubs = "3.14.5"
drf-yasg = "1.21.7"
idna = "3.6"
inflection = "0.5.1"
install = "1.3.5"
jmespath = "1.0.1"
markdown = "3.6"
packaging = "24.0"
psycopg = "3.1.18"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pytz = "2024.1"
pyyaml = "6.0.1"
requests = "2.31.0"
s3transfer = "0.10.1"
setuptools = "69.2.0"
six = "1.16.0"
sqlparse = "0.4.4"
types-awscrt = "0.20.5"
types-pytz = "2024.1.0.20240203"
types-pyyaml = "6.0.12.20240311"
types-requests = "2.31.0.20240311"
types-s3transfer = "0.10.0"
types-stripe = "3.5.2.20240106"
typing-extensions = "4.10.0"
uritemplate = "4.1.1"
urllib3 = "2.2.1"
stripe = "^8.9.0"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
